requirements.txt
streamlit==1.38.0
requests==2.32.3
newspaper3k==0.2.8
beautifulsoup4==4.12.3
transformers==4.44.2
torch==2.3.1
sentencepiece==0.2.0
protobuf==5.27.3
